Enforce strict linting (UP/B/SIM/RUF), add Python matrix CI, remove requirements.txt
Linting:
- Expanded ruff rules: E, F, I, W + UP (pyupgrade), B (bugbear), SIM (simplify), RUF
- Fixed 59 lint violations across 32 files: modernized type hints (Optional→X|None),
  removed unnecessary int(round()), fixed loop variable capture (B023), cleaned up
  unused variables, replaced stdout/stderr=PIPE with capture_output, and more
- All commits are now gated by ruff with 8 rule categories — any push or PR that
  introduces a lint violation will fail the Lint workflow

CI/CD:
- test.yml: Added Python 3.10/3.11/3.12 matrix, pip caching, --tb=short
- Added pytest-cov>=6.0 to test extras for future coverage reporting
- lint.yml unchanged — auto-detects expanded pyproject.toml config

Packaging:
- Removed redundant requirements.txt — pyproject.toml is the single source of truth
  for all dependencies (install via pip install -e '.[test]' or pip install -e '.[dev]')

+++ src/kairos/llm/synopsis/render.py
@@ -28,7 +28,7 @@ def _timecode_to_seconds(timecode: str | None):
         m = int(float(minutes))
         s = float(seconds)
         total = (h * 3600) + (m * 60) + s
-        return int(round(total))
+        return round(total)
     except ValueError:
         return None
